Revert "v1.2.0"
