--- Text/Crawler/.gitignore
@@ -1,5 +1,0 @@
-*.json
-.env
-.DS_Store
-__pycache__/
-__MACOSX/
