--- Text/LLM/model/chatbot/chatbot_constants.py
@@ -1,22 +1,0 @@
-# 카테고리 매핑 상수 (공통 사용)
-label_mapping = {
-    "제로웨이스트": ("ZERO_WASTE", "제로웨이스트"),
-    "플로깅": ("PLOGGING", "플로깅"),
-    "탄소발자국": ("CARBON_FOOTPRINT", "탄소발자국"),
-    "에너지 절약": ("ENERGY_SAVING", "에너지 절약"),
-    "업사이클": ("UPCYCLING", "업사이클"),
-    "문화 공유": ("MEDIA", "문화 공유"),
-    "디지털 탄소": ("DIGITAL_CARBON", "디지털 탄소"),
-    "비건": ("VEGAN", "비건")
-}
-
-# 키워드 및 비속어 필터링 리스트
-ENV_KEYWORDS = [
-    "환경", "지구", "에코", "제로웨이스트", "탄소", "분리수거", "플라스틱", "텀블러", "기후", "친환경",
-    "일회용", "미세먼지", "재활용", "자원", "대중교통", "도보", "비건", "탄소중립", "그린", "에너지",
-    "쓰레기", "아무", "추천", "챌린지", "도움", "도와줘", "자세히", "상세히"
-]
-
-BAD_WORDS = [
-    "시발", "씨발", "fuck", "shit", "개새끼", "병신", "ㅅㅂ", "ㅄ", "ㅂㅅ", "fuckyou", "asshole", "tlqkf", "ㅈ"
-]
