+++ model/chatbot/challenge_docs.txt
