+++ model/chatbot/generate_challenge_docs.py
@@ -1,0 +1,82 @@
+# generate_challenge_docs_unified.py
+# 챌린지 문서 생성기
+import random
+
+def generate_challenge_docs(file_path="challenge_docs.txt", mode="random", num_paragraphs=100):
+    """
+    챌린지 문서를 생성합니다.
+
+    Args:
+        file_path (str): 저장할 파일 경로
+        mode (str): "fixed" (고정 데이터) 또는 "random" (랜덤 조합 데이터)
+        num_paragraphs (int): 랜덤 모드일 때 생성할 문단 수 (default: 100)
+    """
+
+    # 랜덤 데이터 후보 문장
+    random_sentences = [
+           "도시에서 사무직으로 근무하며 점심시간에 제로웨이스트 도시락을 실천하고 있습니다.",
+        "바다 근처에서 영업직으로 일하며 출퇴근 시 플로깅을 실천하고, 지역 커뮤니티에도 확산시키고 있습니다.",
+        "산간 지역에서 재택근무하며 주 1회 가족과 함께 업사이클 공예를 진행하며 친환경 문화를 실천하고 있습니다.",
+        "농촌에서 현장직 근무 중 비닐 대신 천 포대를 활용해 제로웨이스트 농작업을 실현하고 있습니다.",
+        "도시에서 재택근무하며 디지털 탄소 절감을 위해 클라우드 정리와 화면 밝기 조절 캠페인을 진행하고 있습니다.",
+        "현장직 근무자는 매일 차량 대신 자전거로 출퇴근하며 탄소발자국 줄이기에 참여하고 있습니다.",
+        "영업직 근무 중 출장지에서 채식 식당을 탐방하며 비건 챌린지를 SNS에 공유하고 있습니다.",
+        "사무실에서는 냉난방 온도를 2도 낮추는 에너지 절약 캠페인이 진행되고 있습니다.",
+        "재택근무자는 회사 전체에 '이메일 청소의 날'을 제안하고 실행하여 디지털 탄소를 줄이고 있습니다.",
+        "농촌 마을에서 플라스틱 병을 활용한 화분 만들기 워크숍을 통해 업사이클 문화를 확산시키고 있습니다.",
+        "도시 재택근무 팀은 매월 '종이 없는 하루'를 지정해 문서 디지털화를 촉진하고 있습니다.",
+        "산간 지역에서 플로깅 행사를 환경 교육과 결합하여 지역사회 참여를 높이고 있습니다.",
+        "바닷가 인근 회사는 일회용품 제로 캠페인을 운영하며 인증 배지를 수여하고 있습니다.",
+        "사무직 근무자는 비건 점심 도시락 릴레이를 조직해 동료들과 식단 변화를 나누고 있습니다.",
+        "현장 근무 중 음료 캔을 자율 분리수거 코너에 모아 실천을 이어가고 있습니다.",
+        "농촌 재택근무자는 매주 업사이클링 아이디어를 메신저에서 공유하고 실천을 유도하고 있습니다.",
+        "영업직은 고객 방문 시 에코백을 활용해 불필요한 포장을 줄이고 있습니다.",
+        "도시 사무실은 '에코 워킹타임'으로 전등을 1시간 일찍 끄는 에너지 절약을 실천하고 있습니다.",
+        "산간 건설사는 친환경 자재를 활용한 구조물 설계를 추진하고 있습니다.",
+        "바다 인근 영업소는 텀블러 사용자에게 리워드를 제공하는 친환경 마케팅을 운영하고 있습니다.",
+        "도시 사무직으로 근무하며 사무실 내 커피 찌꺼기를 모아 화분 비료로 재활용하고 있습니다.",
+        "산간 지역에서 재택근무하며 화면 밝기 자동 조절 기능을 통해 에너지 절약을 실천하고 있습니다.",
+        "바다 인근에서 영업 활동 중 다회용 텀블러 사용을 생활화하며 쓰레기 배출을 줄이고 있습니다.",
+        "농촌 현장직 근무자는 폐비닐 수거 활동을 정기적으로 실천하고 있습니다.",
+        "도시 재택근무자는 매주 디지털 파일 정리와 클라우드 정리를 함께 수행하고 있습니다.",
+        "영업직은 출장 시 대중교통을 이용하며 탄소배출 감소에 기여하고 있습니다.",
+        "현장직으로 근무하며 점심시간에 자투리 나무를 활용해 소형 가구를 제작하고 있습니다.",
+        "사무직으로 근무하며 매주 한 번 개인 텀블러 세척 캠페인을 진행하고 있습니다.",
+        "바다 인근 재택근무자는 지역 해안 청소 행사에 참여하며 플로깅을 실천하고 있습니다.",
+        "농촌 재택근무자는 계절마다 바뀌는 비건 식단 레시피를 온라인 사내 채널에 공유하고 있습니다.",
+        "도시 사무직은 점심시간마다 텀블러를 사용하며 플라스틱 줄이기에 앞장서고 있습니다.",
+        "산간 지역 재택근무자는 창문을 열고 자연채광을 활용해 전등 사용을 줄이고 있습니다.",
+        "영업직은 출장 중 지하철을 이용하며 이동 중 친환경 콘텐츠를 소비하고 있습니다.",
+        "바닷가 인근 재택근무자는 매일 아침 해변 플로깅을 실천하고 인증하고 있습니다.",
+        "농촌의 현장직은 농기계 공회전을 줄이는 캠페인을 운영 중입니다.",
+        "사무직은 종이 없는 회의를 위해 태블릿을 적극 활용하고 있습니다.",
+        "산간 지역 사무직은 텀블러 사용률을 높이기 위해 매주 세척 캠페인을 운영하고 있습니다.",
+        "도시 재택근무자는 업무 종료 후 콘센트를 뽑는 습관으로 에너지를 절약하고 있습니다.",
+        "현장직은 자투리 목재를 재활용하여 작업 공간을 정비하고 있습니다.",
+        "바다 인근 영업직은 친환경 용기 사용을 고객과 함께 실천하고 있습니다.",
+        "농촌 재택근무자는 일주일에 한 번 지역 농산물만 활용한 식단을 구성하고 있습니다.",
+        "사무직은 클라우드 폴더를 정리하여 디지털 탄소 줄이기 캠페인에 참여하고 있습니다.",
+        "영업직은 다회용 컵을 챙겨 이동하며 일회용 쓰레기를 줄이고 있습니다.",
+        "현장직은 매주 플라스틱 수거량을 기록하여 사내 보고에 활용하고 있습니다.",
+        "산간 재택근무자는 중고 전자기기 교체 캠페인으로 전자폐기물을 줄이고 있습니다.",
+        "도시 사무직은 ‘에코 출근 챌린지’를 통해 도보나 자전거로 출근하는 날을 정하고 있습니다.",
+        "농촌 현장직은 태양광 전등으로 야간 작업 시 에너지 사용을 줄이고 있습니다.",
+        "바다 인근 재택근무자는 친환경 세제를 사용해 세탁 후 오염수를 줄이고 있습니다.",
+        "도시 재택근무자는 프린터 없는 업무 환경을 위해 PDF 문서 활용을 장려하고 있습니다.",
+        "사무직은 종이컵 대신 개인 머그컵 사용을 정착시키기 위한 사내 캠페인을 운영하고 있습니다."
+    ]
+ 
+
+    try:
+        with open(file_path, "w", encoding="utf-8") as f:
+            if mode == "random":
+                for _ in range(num_paragraphs):
+                    num_sentences = random.randint(2, 5)
+                    paragraph = " ".join(random.sample(random_sentences, num_sentences))
+                    f.write(f"{paragraph}\n\n")
+                print(f"랜덤 모드: {num_paragraphs}개 문단 저장 완료")
+            else:
+                print(f"오류: 지원하지 않는 모드입니다. ('fixed' 또는 'random' 사용)")
+
+    except Exception as e:
+        print(f"파일 생성 실패: {str(e)}")
