+++ model/verify/event_challenge_prompt.py
