+++ model/verify/worker.py
@@ -22,7 +22,6 @@ def run_worker():
         print("callback 메시지 수신됨", message)
         try:
             data = json.loads(message.data.decode("utf-8"))
-
             blob_name = data["imageUrl"].split("/")[-1]
             challenge_type = data["type"]
             challenge_id = int(data["challengeId"])
@@ -39,9 +38,9 @@ def run_worker():
 
             # 콜백 URL 내 challengeId 치환
             # -> CALLBACK_URL에 {verificationId}가 포함되는 경우, Python에서 실제 전송 전에 .format() 또는 f-string으로 치환해줘야함 
-            formatted_url = os.getenv("CALLBACK_URL_VERIFY").format(verificationId=data["verificationId"])
+            formatted_url = os.getenv("CALLBACK_URL").format(verificationId=data["verificationId"])
 
-            
+
             # 결과 콜백 전송
             requests.post(formatted_url, json={
                 "type": data["type"],
@@ -50,7 +49,7 @@ def run_worker():
                 "date": data["date"],
                 "result": is_verified
             })
-            
+
             message.ack()
 
         except Exception as e:
