+++ router/censorship_router.py
@@ -6,7 +6,7 @@ from starlette.status import HTTP_422_UNPROCESSABLE_ENTITY
 from pydantic import BaseModel, Field
 from typing import List, Optional
 
-from ..model.censorship.LLM_censorship_model import CensorshipModel
+from model.censorship.LLM_censorship_model import CensorshipModel
 
 router = APIRouter()
 model = CensorshipModel()
