+++ test_ai_chatbot.http
@@ -3,10 +3,9 @@ POST http://localhost:9000/ai/chatbot/recommendation/base-info
 Content-Type: application/json
 
 {
-  "sessionId": "user123",
   "location": "도시",
   "workType": "사무직",
-  "category": "제로웨이스트"
+  "category": "비건"
 }
 
 ###
@@ -16,6 +15,5 @@ POST http://localhost:9000/ai/chatbot/recommendation/free-text
 Content-Type: application/json
 
 {
-  "sessionId": "user123",
-  "message": "아무거나 추천"
+  "message": "비건 관련 챌린지 추천"
 }
